refactor ref -> var
